Split out MethodExtensionResult
